Break core chunk code into a module
